requirements.txt
flask==2.3.3
flask-cors==4.0.0
groq==0.4.2
requests==2.31.0
python-dotenv==1.0.0
elevenlabs==0.2.26
transformers==4.36.2
torch==2.1.2
accelerate==0.25.0
tensorflow==2.15.0
keras==2.15.0